+++ CHANGELOG.md
@@ -3,6 +3,14 @@
 The format is based on and uses the types of changes according to [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [2.7.2] - 2023-06-15
+
+### Changed
+
+- SPSWakeUP.ps1
+ - BREAKING CHANGE: Remove Add-HostsEntry function
+ - Using the CmdLet Add-Content for update HOSTS file
+
 ## [2.7.1] - 2023-05-10
 
 ### Changed